requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python-headless
Pillow
python-multipart
